timecard and Timecard functions
